# api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ACCESS Ready API - Adaptive ELL exit prep for K-12 students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teachers/{teacherId}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get teacher by ID
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    patch:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update teacher profile
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeacherBody"
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "404":
          description: Teacher not found
  /teachers/{teacherId}/students:
    get:
      operationId: listStudents
      tags: [teachers]
      summary: List all students for a teacher
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of students with current levels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentWithProgress"
  /teachers/{teacherId}/dashboard:
    get:
      operationId: getTeacherDashboard
      tags: [dashboard]
      summary: Get teacher dashboard summary
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary with alerts and watch lists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherDashboard"
  /teachers/{teacherId}/export:
    get:
      operationId: exportStudentsCsv
      tags: [dashboard]
      summary: Export student data as CSV
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
...
components:
  schemas:
...
    UpdateTeacherBody:
      type: object
      properties:
        name:
          type: string
        school:
          type: ["string", "null"]
          description: Free-text school name (legacy). Prefer schoolId when linking to a structured school record.
        schoolId:
          type: ["string", "null"]
          description: UUID of a school in the schools table. Setting this links the teacher to that school.
        avatarUrl:
          type: ["string", "null"]
    Teacher:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        school:
          type: ["string", "null"]
          description: Legacy free-text school name.
        schoolId:
          type: ["string", "null"]
          description: FK to the schools table. Null if teacher is not linked to a school.
        schoolName:
          type: ["string", "null"]
          description: Resolved school name from the schools table (joined).
        districtId:
          type: ["string", "null"]
          description: District ID via the linked school. Null if no school linked.
        isSolo:
          type: boolean
          description: >
            True when the teacher is independent (no schoolId).
            Solo teachers manage billing and their own student roster.
            School-managed teachers are under a principal and cannot add students.
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - isSolo
        - createdAt
...
    Student:
      type: object
      properties:
        id:
          type: string
        teacherId:
          type: string
        name:
          type: string
        gradeBand:
          type: string
        stateAssessment:
          type: string
        homeLanguage:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        currentStreak:
          type: integer
        streakShieldAvailable:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - teacherId
        - name
        - gradeBand
        - stateAssessment
        - currentStreak
        - streakShieldAvailable
        - createdAt
    DomainLevel:
      type: object
      properties:
        domain:
          type: string
          enum: [listening, speaking, reading, writing]
        currentLevel:
          type: number
        exitThreshold:
          type: number
        gap:
          type: number
        normalizedLevel:
          type: number
        levelLabel:
          type: string
        atExit:
          type: boolean
        lastSessionScore:
          type: ["integer", "null"]
        lastPracticed:
          type: ["string", "null"]
      required:
        - domain
        - currentLevel
        - exitThreshold
        - gap
        - normalizedLevel
        - levelLabel
        - atExit
...
    StudentWithProgress:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
        domains:
          type: array
          items:
            $ref: "#/components/schemas/DomainLevel"
        lastSessionDate:
          type: ["string", "null"]
        isInactive:
          type: boolean
        isActive:
          type: boolean
        isNearExit:
          type: boolean
        isStalled:
          type: boolean
        stalledDomain:
          type: ["string", "null"]
        teacherRecommendation:
          type: ["string", "null"]
      required:
        - student
        - domains
        - isInactive
        - isActive
        - isNearExit
        - isStalled
...
    TeacherDashboard:
      type: object
      properties:
        teacherId:
          type: string
        totalStudents:
          type: integer
        activeToday:
          type: integer
        exitWatchList:
          type: array
          items:
            $ref: "#/components/schemas/StudentWithProgress"
        stalledStudents:
          type: array
          items:
            $ref: "#/components/schemas/StudentWithProgress"
        inactiveStudents:
          type: array
          items:
            $ref: "#/components/schemas/StudentWithProgress"
        allStudents:
          type: array
          items:
            $ref: "#/components/schemas/StudentWithProgress"
      required:
        - teacherId
        - totalStudents
        - activeToday
        - exitWatchList
        - stalledStudents
        - inactiveStudents
        - allStudents
...